requirements.txt
requests>=2.31.0
psycopg2-binary>=2.9.7
python-dotenv>=1.0.0
flask>=2.3.0
psutil>=5.9.0

# Security packages
flask-login>=0.6.3
flask-session>=0.5.0
flask-wtf>=1.2.1
flask-limiter>=3.5.0
flask-talisman>=1.1.0
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
requests-oauthlib>=1.3.1
pyopenssl>=23.3.0
cryptography>=41.0.7
redis>=5.0.1  # For session storage and rate limiting 